add man support
