+++ clidocstool_man_test.go
@@ -1,4 +1,4 @@
-// Copyright 2017 cli-docs-tool authors
+// Copyright 2024 cli-docs-tool authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -17,33 +17,50 @@ package clidocstool
 import (
 	"io/fs"
 	"os"
+	"path"
 	"path/filepath"
-	"strings"
+	"regexp"
+	"strconv"
 	"testing"
+	"time"
 
+	"github.com/spf13/cobra/doc"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
 //nolint:errcheck
-func TestGenYamlTree(t *testing.T) {
+func TestGenManTree(t *testing.T) {
 	setup()
 	tmpdir := t.TempDir()
+
+	epoch, err := time.Parse("2006-Jan-02", "2020-Jan-10")
+	require.NoError(t, err)
+	t.Setenv("SOURCE_DATE_EPOCH", strconv.FormatInt(epoch.Unix(), 10))
+
+	require.NoError(t, copyFile(path.Join("fixtures", "buildx_stop.pre.md"), path.Join(tmpdir, "buildx_stop.md")))
 
 	c, err := New(Options{
 		Root:      buildxCmd,
 		SourceDir: tmpdir,
 		Plugin:    true,
+		ManHeader: &doc.GenManHeader{
+			Title:   "DOCKER",
+			Section: "1",
+			Source:  "Docker Community",
+			Manual:  "Docker User Manuals",
+		},
 	})
 	require.NoError(t, err)
-	require.NoError(t, c.GenYamlTree(buildxCmd))
+	require.NoError(t, c.GenManTree(buildxCmd))
 
 	seen := make(map[string]struct{})
+	remanpage := regexp.MustCompile(`\.\d+$`)
 
 	filepath.Walk("fixtures", func(path string, info fs.FileInfo, err error) error {
 		fname := filepath.Base(path)
-		// ignore dirs and any file that is not a .yaml file
-		if info.IsDir() || !strings.HasSuffix(fname, ".yaml") {
+		// ignore dirs and any file that is not a manpage
+		if info.IsDir() || !remanpage.MatchString(fname) {
 			return nil
 		}
 		t.Run(fname, func(t *testing.T) {
@@ -62,8 +79,8 @@ func TestGenYamlTree(t *testing.T) {
 
 	filepath.Walk(tmpdir, func(path string, info fs.FileInfo, err error) error {
 		fname := filepath.Base(path)
-		// ignore dirs and any file that is not a .yaml file
-		if info.IsDir() || !strings.HasSuffix(fname, ".yaml") {
+		// ignore dirs and any file that is not a manpage
+		if info.IsDir() || !remanpage.MatchString(fname) {
 			return nil
 		}
 		t.Run("seen_"+fname, func(t *testing.T) {
