+++ clidocstool_md_test.go
@@ -27,6 +27,7 @@ import (
 
 //nolint:errcheck
 func TestGenMarkdownTree(t *testing.T) {
+	setup()
 	tmpdir := t.TempDir()
 
 	require.NoError(t, copyFile(path.Join("fixtures", "buildx_stop.pre.md"), path.Join(tmpdir, "buildx_stop.md")))
