+++ fixtures/docker-buildx-build.1
@@ -1,0 +1,122 @@
+.nh
+.TH "DOCKER" "1" "Jan 2020" "Docker Community" "Docker User Manuals"
+
+.SH NAME
+.PP
+docker-buildx-build - Start a build
+
+
+.SH SYNOPSIS
+.PP
+\fBdocker buildx build [OPTIONS] PATH | URL | -\fP
+
+
+.SH DESCRIPTION
+.PP
+Start a build
+
+
+.SH OPTIONS
+.PP
+\fB--add-host\fP=[]
+	Add a custom host-to-IP mapping (format: 'host:ip')
+
+.PP
+\fB--allow\fP=[]
+	Allow extra privileged entitlement (e.g., "network.host", "security.insecure")
+
+.PP
+\fB--build-arg\fP=[]
+	Set build-time variables
+
+.PP
+\fB--cache-from\fP=[]
+	External cache sources (e.g., "user/app:cache", "type=local,src=path/to/dir")
+
+.PP
+\fB--cache-to\fP=[]
+	Cache export destinations (e.g., "user/app:cache", "type=local,dest=path/to/dir")
+
+.PP
+\fB--cgroup-parent\fP=""
+	Optional parent cgroup for the container
+
+.PP
+\fB--detach\fP[=true]
+	Dummy flag that tests boolean flags with true as default
+
+.PP
+\fB-f\fP, \fB--file\fP=""
+	Name of the Dockerfile (default: "PATH/Dockerfile")
+
+.PP
+\fB-h\fP, \fB--help\fP[=false]
+	help for build
+
+.PP
+\fB--iidfile\fP=""
+	Write the image ID to the file
+
+.PP
+\fB--label\fP=[]
+	Set metadata for an image
+
+.PP
+\fB--load\fP[=false]
+	Shorthand for "--output=type=docker"
+
+.PP
+\fB--network\fP="default"
+	Set the networking mode for the "RUN" instructions during build
+
+.PP
+\fB-o\fP, \fB--output\fP=[]
+	Output destination (format: "type=local,dest=path")
+
+.PP
+\fB--platform\fP=[]
+	Set target platform for build
+
+.PP
+\fB--push\fP[=false]
+	Shorthand for "--output=type=registry"
+
+.PP
+\fB-q\fP, \fB--quiet\fP[=false]
+	Suppress the build output and print image ID on success
+
+.PP
+\fB--secret\fP=[]
+	Secret file to expose to the build (format: "id=mysecret,src=/local/secret")
+
+.PP
+\fB--shm-size\fP=""
+	Size of "/dev/shm"
+
+.PP
+\fB--ssh\fP=[]
+	SSH agent socket or keys to expose to the build
+format: "default|[=|[,]]"
+
+.PP
+\fB-t\fP, \fB--tag\fP=[]
+	Name and optionally a tag (format: "name:tag")
+
+.PP
+\fB--target\fP=""
+	Set the target build stage to build.
+
+.PP
+\fB--ulimit\fP=""
+	Ulimit options
+
+
+.SH OPTIONS INHERITED FROM PARENT COMMANDS
+.PP
+\fB--builder\fP=""
+	Override the configured builder instance
+
+
+.SH SEE ALSO
+.PP
+\fBdocker-buildx(1)\fP
