+++ fixtures/docker-buildx-install.1
@@ -1,0 +1,33 @@
+.nh
+.TH "DOCKER" "1" "Jan 2020" "Docker Community" "Docker User Manuals"
+
+.SH NAME
+.PP
+docker-buildx-install - Install buildx as a 'docker builder' alias
+
+
+.SH SYNOPSIS
+.PP
+\fBdocker buildx install\fP
+
+
+.SH DESCRIPTION
+.PP
+Install buildx as a 'docker builder' alias
+
+
+.SH OPTIONS
+.PP
+\fB-h\fP, \fB--help\fP[=false]
+	help for install
+
+
+.SH OPTIONS INHERITED FROM PARENT COMMANDS
+.PP
+\fB--builder\fP=""
+	Override the configured builder instance
+
+
+.SH SEE ALSO
+.PP
+\fBdocker-buildx(1)\fP
