+++ fixtures/docker-buildx-stop.1
@@ -1,0 +1,33 @@
+.nh
+.TH "DOCKER" "1" "Jan 2020" "Docker Community" "Docker User Manuals"
+
+.SH NAME
+.PP
+docker-buildx-stop - Stop builder instance
+
+
+.SH SYNOPSIS
+.PP
+\fBdocker buildx stop [NAME]\fP
+
+
+.SH DESCRIPTION
+.PP
+Stop builder instance
+
+
+.SH OPTIONS
+.PP
+\fB-h\fP, \fB--help\fP[=false]
+	help for stop
+
+
+.SH OPTIONS INHERITED FROM PARENT COMMANDS
+.PP
+\fB--builder\fP=""
+	Override the configured builder instance
+
+
+.SH SEE ALSO
+.PP
+\fBdocker-buildx(1)\fP
