+++ fixtures/docker-buildx.1
@@ -1,0 +1,31 @@
+.nh
+.TH "DOCKER" "1" "Jan 2020" "Docker Community" "Docker User Manuals"
+
+.SH NAME
+.PP
+docker-buildx - Docker Buildx
+
+
+.SH SYNOPSIS
+.PP
+\fBdocker buildx\fP
+
+
+.SH DESCRIPTION
+.PP
+Extended build capabilities with BuildKit
+
+
+.SH OPTIONS
+.PP
+\fB--builder\fP=""
+	Override the configured builder instance
+
+.PP
+\fB-h\fP, \fB--help\fP[=false]
+	help for buildx
+
+
+.SH SEE ALSO
+.PP
+\fBdocker(1)\fP, \fBdocker-buildx-build(1)\fP, \fBdocker-buildx-stop(1)\fP
